+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/EpisodeManager.kt
@@ -165,4 +165,6 @@ interface EpisodeManager {
         toEpochMsCurrentYear: Long,
     ): YearOverYearListeningTime
     suspend fun countEpisodesStartedAndCompleted(fromEpochMs: Long, toEpochMs: Long): EpisodesStartedAndCompleted
+
+    suspend fun updateDownloadUrl(episode: PodcastEpisode): String?
 }